Migrates Ruby version to 1.8.7 and adds debug flag for printing.

+++ src/main.rb
@@ -3,9 +3,21 @@
 # Author: Jake Wilson
 # Date: 08/29/15
 # 
+#
 
-require_relative 'scanner.rb'
-require_relative 'types.rb'
+$debug = false
+
+require "./scanner.rb"
+require "./types.rb"
+
+if !File.exists? ARGV[0]
+    puts ARGV[0] + " does not exist"
+    exit 1
+end
+
+if ARGV.count >= 2
+    $debug = true
+end
 
 input = File.open(ARGV[0], 'r')
 
@@ -13,4 +25,5 @@ Scanner.entire_file = input.readlines
 input.rewind
 
 while (ret = Scanner.getTokenType(input)) != TokenType::EOF
+
 end
